final-pjt-back/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.1.2
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.8
drf_swagger==0.1.3
environ==1.0
frozenlist==1.5.0
greenlet==0.4.13
hpy==0.9.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
maturin==1.7.4
multidict==6.1.0
oauthlib==3.2.2
openai==0.27.8
packaging==24.2
pillow==11.0.0
propcache==0.2.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
readline==6.2.4.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
shell==1.0.1
six==1.16.0
sqlparse==0.5.1
tomli==2.1.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
utils==1.0.2
xmltodict==0.14.2
yarl==1.18.0
